Move ThemeToggle to component folder
