+++ src/components/layout.js
@@ -4,7 +4,7 @@ import { graphql, Link, useStaticQuery } from "gatsby"
 import SEO from "./SEO"
 import * as PropTypes from "prop-types"
 import BackToTop from "./BackToTop/BackToTop"
-import ThemeToggle from "./theme-toggle"
+import ThemeToggle from "./ThemeToggle/ThemeToggle"
 import Footer from "./footer"
 
 const Layout = ({ tabTitle, pageTitle, pagination, tags, home, children, description, backToTop }) => {
