+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import style from "./index.module.css"
 import { Link } from "gatsby"
 import SEO from "../components/SEO"
-import ThemeToggle from "../components/theme-toggle"
+import ThemeToggle from "../components/ThemeToggle/ThemeToggle"
 import { graphql } from "gatsby"
 import profilePicture from "./assets/profile_picture.png"
 import { AiFillLinkedin, AiFillGithub, AiFillTwitterSquare } from "react-icons/all"
